# pkg/ucp/api/v20231001preview/zz_generated_kubernetesplanes_client.go
//go:build go1.18
// +build go1.18

// Licensed under the Apache License, Version 2.0 . See LICENSE in the repository root for license information.
// Code generated by Microsoft (R) AutoRest Code Generator. DO NOT EDIT.
// Changes may cause incorrect behavior and will be lost if the code is regenerated.

package v20231001preview

import (
	"context"
	"errors"
	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
	"net/http"
	"net/url"
	"strings"
)

// KubernetesPlanesClient contains the methods for the KubernetesPlanes group.
// Don't use this type directly, use NewKubernetesPlanesClient() instead.
type KubernetesPlanesClient struct {
	internal *arm.Client
}

// NewKubernetesPlanesClient creates a new instance of KubernetesPlanesClient with the specified values.
//   - credential - used to authorize requests. Usually a credential from azidentity.
//   - options - pass nil to accept the default values.
func NewKubernetesPlanesClient(credential azcore.TokenCredential, options *arm.ClientOptions) (*KubernetesPlanesClient, error) {
	cl, err := arm.NewClient(moduleName+".KubernetesPlanesClient", moduleVersion, credential, options)
	if err != nil {
		return nil, err
	}
	client := &KubernetesPlanesClient{
	internal: cl,
	}
	return client, nil
}

// BeginCreateOrUpdate - Create or update a plane
// If the operation fails it returns an *azcore.ResponseError type.
//
// Generated from API version 2023-10-01-preview
//   - planeName - The plane name.
//   - resource - Resource create parameters.
//   - options - KubernetesPlanesClientBeginCreateOrUpdateOptions contains the optional parameters for the KubernetesPlanesClient.BeginCreateOrUpdate
//     method.
func (client *KubernetesPlanesClient) BeginCreateOrUpdate(ctx context.Context, planeName string, resource KubernetesPlaneResource, options *KubernetesPlanesClientBeginCreateOrUpdateOptions) (*runtime.Poller[KubernetesPlanesClientCreateOrUpdateResponse], error) {
	if options == nil || options.ResumeToken == "" {
		resp, err := client.createOrUpdate(ctx, planeName, resource, options)
		if err != nil {
			return nil, err
		}
		poller, err := runtime.NewPoller(resp, client.internal.Pipeline(), &runtime.NewPollerOptions[KubernetesPlanesClientCreateOrUpdateResponse]{
			FinalStateVia: runtime.FinalStateViaAzureAsyncOp,
		})
		return poller, err
	} else {
		return runtime.NewPollerFromResumeToken[KubernetesPlanesClientCreateOrUpdateResponse](options.ResumeToken, client.internal.Pipeline(), nil)
	}
}

// CreateOrUpdate - Create or update a plane
// If the operation fails it returns an *azcore.ResponseError type.
//
// Generated from API version 2023-10-01-preview
func (client *KubernetesPlanesClient) createOrUpdate(ctx context.Context, planeName string, resource KubernetesPlaneResource, options *KubernetesPlanesClientBeginCreateOrUpdateOptions) (*http.Response, error) {
	var err error
	req, err := client.createOrUpdateCreateRequest(ctx, planeName, resource, options)
	if err != nil {
		return nil, err
	}
	httpResp, err := client.internal.Pipeline().Do(req)
	if err != nil {
		return nil, err
	}
	if !runtime.HasStatusCode(httpResp, http.StatusOK, http.StatusCreated) {
		err = runtime.NewResponseError(httpResp)
		return nil, err
	}
	return httpResp, nil
}

// createOrUpdateCreateRequest creates the CreateOrUpdate request.
func (client *KubernetesPlanesClient) createOrUpdateCreateRequest(ctx context.Context, planeName string, resource KubernetesPlaneResource, options *KubernetesPlanesClientBeginCreateOrUpdateOptions) (*policy.Request, error) {
	urlPath := "/planes/kubernetes/{planeName}"
	if planeName == "" {
		return nil, errors.New("parameter planeName cannot be empty")
	}
	urlPath = strings.ReplaceAll(urlPath, "{planeName}", url.PathEscape(planeName))
	req, err := runtime.NewRequest(ctx, http.MethodPut, runtime.JoinPaths(client.internal.Endpoint(), urlPath))
	if err != nil {
		return nil, err
	}
	reqQP := req.Raw().URL.Query()
	reqQP.Set("api-version", "2023-10-01-preview")
	req.Raw().URL.RawQuery = reqQP.Encode()
	req.Raw().Header["Accept"] = []string{"application/json"}
	if err := runtime.MarshalAsJSON(req, resource); err != nil {
	return nil, err
}
	return req, nil
}

// BeginDelete - Delete a plane
// If the operation fails it returns an *azcore.ResponseError type.
//
// Generated from API version 2023-10-01-preview
//   - planeName - The plane name.
//   - options - KubernetesPlanesClientBeginDeleteOptions contains the optional parameters for the KubernetesPlanesClient.BeginDelete
//     method.
func (client *KubernetesPlanesClient) BeginDelete(ctx context.Context, planeName string, options *KubernetesPlanesClientBeginDeleteOptions) (*runtime.Poller[KubernetesPlanesClientDeleteResponse], error) {
	if options == nil || options.ResumeToken == "" {
		resp, err := client.deleteOperation(ctx, planeName, options)
		if err != nil {
			return nil, err
		}
		poller, err := runtime.NewPoller(resp, client.internal.Pipeline(), &runtime.NewPollerOptions[KubernetesPlanesClientDeleteResponse]{
			FinalStateVia: runtime.FinalStateViaLocation,
		})
		return poller, err
	} else {
		return runtime.NewPollerFromResumeToken[KubernetesPlanesClientDeleteResponse](options.ResumeToken, client.internal.Pipeline(), nil)
	}
}

// Delete - Delete a plane
// If the operation fails it returns an *azcore.ResponseError type.
//
// Generated from API version 2023-10-01-preview
func (client *KubernetesPlanesClient) deleteOperation(ctx context.Context, planeName string, options *KubernetesPlanesClientBeginDeleteOptions) (*http.Response, error) {
	var err error
	req, err := client.deleteCreateRequest(ctx, planeName, options)
	if err != nil {
		return nil, err
	}
	httpResp, err := client.internal.Pipeline().Do(req)
	if err != nil {
		return nil, err
	}
	if !runtime.HasStatusCode(httpResp, http.StatusOK, http.StatusAccepted, http.StatusNoContent) {
		err = runtime.NewResponseError(httpResp)
		return nil, err
	}
	return httpResp, nil
}

// deleteCreateRequest creates the Delete request.
func (client *KubernetesPlanesClient) deleteCreateRequest(ctx context.Context, planeName string, options *KubernetesPlanesClientBeginDeleteOptions) (*policy.Request, error) {
	urlPath := "/planes/kubernetes/{planeName}"
	if planeName == "" {
		return nil, errors.New("parameter planeName cannot be empty")
	}
	urlPath = strings.ReplaceAll(urlPath, "{planeName}", url.PathEscape(planeName))
	req, err := runtime.NewRequest(ctx, http.MethodDelete, runtime.JoinPaths(client.internal.Endpoint(), urlPath))
	if err != nil {
		return nil, err
	}
	reqQP := req.Raw().URL.Query()
	reqQP.Set("api-version", "2023-10-01-preview")
	req.Raw().URL.RawQuery = reqQP.Encode()
	req.Raw().Header["Accept"] = []string{"application/json"}
	return req, nil
}

// Get - Get a plane by name
// If the operation fails it returns an *azcore.ResponseError type.
//
// Generated from API version 2023-10-01-preview
//   - planeName - The plane name.
//   - options - KubernetesPlanesClientGetOptions contains the optional parameters for the KubernetesPlanesClient.Get method.
func (client *KubernetesPlanesClient) Get(ctx context.Context, planeName string, options *KubernetesPlanesClientGetOptions) (KubernetesPlanesClientGetResponse, error) {
	var err error
	req, err := client.getCreateRequest(ctx, planeName, options)
	if err != nil {
		return KubernetesPlanesClientGetResponse{}, err
	}
	httpResp, err := client.internal.Pipeline().Do(req)
	if err != nil {
		return KubernetesPlanesClientGetResponse{}, err
	}
	if !runtime.HasStatusCode(httpResp, http.StatusOK) {
		err = runtime.NewResponseError(httpResp)
		return KubernetesPlanesClientGetResponse{}, err
	}
	resp, err := client.getHandleResponse(httpResp)
	return resp, err
}

// getCreateRequest creates the Get request.
func (client *KubernetesPlanesClient) getCreateRequest(ctx context.Context, planeName string, options *KubernetesPlanesClientGetOptions) (*policy.Request, error) {
	urlPath := "/planes/kubernetes/{planeName}"
	if planeName == "" {
		return nil, errors.New("parameter planeName cannot be empty")
	}
	urlPath = strings.ReplaceAll(urlPath, "{planeName}", url.PathEscape(planeName))
	req, err := runtime.NewRequest(ctx, http.MethodGet, runtime.JoinPaths(client.internal.Endpoint(), urlPath))
	if err != nil {
		return nil, err
	}
	reqQP := req.Raw().URL.Query()
	reqQP.Set("api-version", "2023-10-01-preview")
	req.Raw().URL.RawQuery = reqQP.Encode()
	req.Raw().Header["Accept"] = []string{"application/json"}
	return req, nil
}

// getHandleResponse handles the Get response.
func (client *KubernetesPlanesClient) getHandleResponse(resp *http.Response) (KubernetesPlanesClientGetResponse, error) {
	result := KubernetesPlanesClientGetResponse{}
	if err := runtime.UnmarshalAsJSON(resp, &result.KubernetesPlaneResource); err != nil {
		return KubernetesPlanesClientGetResponse{}, err
	}
	return result, nil
}

// NewListPager - List kubernetes planes
//
// Generated from API version 2023-10-01-preview
//   - options - KubernetesPlanesClientListOptions contains the optional parameters for the KubernetesPlanesClient.NewListPager
//     method.
func (client *KubernetesPlanesClient) NewListPager(options *KubernetesPlanesClientListOptions) (*runtime.Pager[KubernetesPlanesClientListResponse]) {
	return runtime.NewPager(runtime.PagingHandler[KubernetesPlanesClientListResponse]{
		More: func(page KubernetesPlanesClientListResponse) bool {
			return page.NextLink != nil && len(*page.NextLink) > 0
		},
		Fetcher: func(ctx context.Context, page *KubernetesPlanesClientListResponse) (KubernetesPlanesClientListResponse, error) {
			var req *policy.Request
			var err error
			if page == nil {
				req, err = client.listCreateRequest(ctx, options)
			} else {
				req, err = runtime.NewRequest(ctx, http.MethodGet, *page.NextLink)
			}
			if err != nil {
				return KubernetesPlanesClientListResponse{}, err
			}
			resp, err := client.internal.Pipeline().Do(req)
			if err != nil {
				return KubernetesPlanesClientListResponse{}, err
			}
			if !runtime.HasStatusCode(resp, http.StatusOK) {
				return KubernetesPlanesClientListResponse{}, runtime.NewResponseError(resp)
			}
			return client.listHandleResponse(resp)
		},
	})
}

// listCreateRequest creates the List request.
func (client *KubernetesPlanesClient) listCreateRequest(ctx context.Context, options *KubernetesPlanesClientListOptions) (*policy.Request, error) {
	urlPath := "/planes/kubernetes"
	req, err := runtime.NewRequest(ctx, http.MethodGet, runtime.JoinPaths(client.internal.Endpoint(), urlPath))
	if err != nil {
		return nil, err
	}
	reqQP := req.Raw().URL.Query()
	reqQP.Set("api-version", "2023-10-01-preview")
	req.Raw().URL.RawQuery = reqQP.Encode()
	req.Raw().Header["Accept"] = []string{"application/json"}
	return req, nil
}

// listHandleResponse handles the List response.
func (client *KubernetesPlanesClient) listHandleResponse(resp *http.Response) (KubernetesPlanesClientListResponse, error) {
	result := KubernetesPlanesClientListResponse{}
	if err := runtime.UnmarshalAsJSON(resp, &result.KubernetesPlaneResourceListResult); err != nil {
		return KubernetesPlanesClientListResponse{}, err
	}
	return result, nil
}

// BeginUpdate - Update a plane
// If the operation fails it returns an *azcore.ResponseError type.
//
// Generated from API version 2023-10-01-preview
//   - planeName - The plane name.
//   - properties - The resource properties to be updated.
//   - options - KubernetesPlanesClientBeginUpdateOptions contains the optional parameters for the KubernetesPlanesClient.BeginUpdate
//     method.
func (client *KubernetesPlanesClient) BeginUpdate(ctx context.Context, planeName string, properties KubernetesPlaneResourceTagsUpdate, options *KubernetesPlanesClientBeginUpdateOptions) (*runtime.Poller[KubernetesPlanesClientUpdateResponse], error) {
	if options == nil || options.ResumeToken == "" {
		resp, err := client.update(ctx, planeName, properties, options)
		if err != nil {
			return nil, err
		}
		poller, err := runtime.NewPoller(resp, client.internal.Pipeline(), &runtime.NewPollerOptions[KubernetesPlanesClientUpdateResponse]{
			FinalStateVia: runtime.FinalStateViaLocation,
		})
		return poller, err
	} else {
		return runtime.NewPollerFromResumeToken[KubernetesPlanesClientUpdateResponse](options.ResumeToken, client.internal.Pipeline(), nil)
	}
}

// Update - Update a plane
// If the operation fails it returns an *azcore.ResponseError type.
//
// Generated from API version 2023-10-01-preview
func (client *KubernetesPlanesClient) update(ctx context.Context, planeName string, properties KubernetesPlaneResourceTagsUpdate, options *KubernetesPlanesClientBeginUpdateOptions) (*http.Response, error) {
	var err error
	req, err := client.updateCreateRequest(ctx, planeName, properties, options)
	if err != nil {
		return nil, err
	}
	httpResp, err := client.internal.Pipeline().Do(req)
	if err != nil {
		return nil, err
	}
	if !runtime.HasStatusCode(httpResp, http.StatusOK, http.StatusAccepted) {
		err = runtime.NewResponseError(httpResp)
		return nil, err
	}
	return httpResp, nil
}

// updateCreateRequest creates the Update request.
func (client *KubernetesPlanesClient) updateCreateRequest(ctx context.Context, planeName string, properties KubernetesPlaneResourceTagsUpdate, options *KubernetesPlanesClientBeginUpdateOptions) (*policy.Request, error) {
	urlPath := "/planes/kubernetes/{planeName}"
	if planeName == "" {
		return nil, errors.New("parameter planeName cannot be empty")
	}
	urlPath = strings.ReplaceAll(urlPath, "{planeName}", url.PathEscape(planeName))
	req, err := runtime.NewRequest(ctx, http.MethodPatch, runtime.JoinPaths(client.internal.Endpoint(), urlPath))
	if err != nil {
		return nil, err
	}
	reqQP := req.Raw().URL.Query()
	reqQP.Set("api-version", "2023-10-01-preview")
	req.Raw().URL.RawQuery = reqQP.Encode()
	req.Raw().Header["Accept"] = []string{"application/json"}
	if err := runtime.MarshalAsJSON(req, properties); err != nil {
	return nil, err
}
	return req, nil
}